Clean up new post interface

+++ ember/app/controllers/posts/edit.js
@@ -7,10 +7,12 @@ var EditController = Ember.ObjectController.extend({
     this.autoSave();
   },
 
-  autoSave: function () {
-    Ember.run.later(this, function() {
+  autoSave: function() {
+    this.timer = Ember.run.later(this, function() {
+      console.log("hello");
       if(this.get("isDirty")){
-        var alert = Em.$(".alert");
+        var alert = Ember.$(".alert");
+        var notificationMessage = 'Your post "' + this.get("title") + ' was auto saved';
 
         this.model.save().catch(function(reason){
           if(reason.status === 500){
@@ -18,41 +20,78 @@ var EditController = Ember.ObjectController.extend({
           }
         });
 
-        alert.text("Your post was auto saved");
+        alert.text(notificationMessage).show();
+        if(document.hidden){
+          this.desktopNotifcation(notificationMessage);
+        }
 
         window.setTimeout(function(){
-          alert.text("");
-        }, 6000);
+          alert.text("").hide();
+        }, 5000);
+
       }
       this.autoSave();
-    }, 120000); //120000 = 2 mins
+    }, 60000); //60000 = 1 min
+  },
+
+  stopAutoSave: function(){
+    Ember.run.cancel(this.timer);
+  },
+
+  desktopNotifcation: function(message) {
+    if(window.Notification.permission === "granted") {
+      new window.Notification(message);
+    }
+    else if(window.Notification.permission !== "denied") {
+      window.Notification.requestPermission(function (permission) {
+        if(permission === "granted") {
+          new window.Notification(message);
+        }
+      });
+    }
   },
 
   actions: {
 
     destroy: function() {
-      this.store.find('posts', this.model.id).then(function (post) {
-        post.destroyRecord();
-      });
-      return this.transitionTo('dashboard');
+      var prompt = window.confirm("Are you sure you want to delete this?");
+      if(prompt) {
+        this.store.find("posts", this.model.id).then(function (post) {
+          post.destroyRecord();
+        });
+        return this.transitionTo("dashboard");
+      }
     },
 
     save: function() {
       return this.model.save().then((function(_this) {
         return function() {
           if(_this.model._data.is_published === true){
-            return _this.transitionToRoute('posts.show', _this.model);
+            return _this.transitionToRoute("posts.show", _this.model);
+          } else {
+            var notificationMessage = "Your post was saved";
+            var alert = Ember.$(".alert");
+
+            window.setTimeout(function(){
+              alert.text("").hide();
+            }, 5000);
+            alert.text(notificationMessage).show();
           }
         };
       })(this));
     },
 
     cancel: function() {
-      if (this.model.isDirty) {
+      if(this.model.isDirty) {
         this.model.rollback();
       }
-      return this.transitionTo('posts.show', this.model);
+      return this.transitionTo("posts.show", this.model);
     },
+
+    togglePreview: function(){
+      Ember.$(".preview").toggleClass("hide");
+    }
+
   }
 
 });
